+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Allow `ArrayLiteral` opening brackets to have trailing comments.
 - Allow different line suffix nodes to have different priorities.
 - Better support for encoding by properly reading encoding magic comments.
+- Support singleton single-line method definitions.
 
 ### Changed